requirements.txt
langchain==0.3.19
langchain-openai==0.3.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1